requirements.txt
numpy~=1.26.0
scikit-learn~=1.3.1
seaborn~=0.13.2
matplotlib~=3.8.0
PyYAML~=6.0.1
pandas~=2.1.1
SciencePlots~=2.1.1
scipy~=1.11.3
tqdm~=4.66.1
plotly~=5.20.0
torch~=2.1.2
lightning~=2.1.2
statsmodels~=0.14.0
pygam~=0.9.0
PuLP~=2.7.0